criada e alteradas funções. input de arquivo.
• adicionado o arquivo 'GeradorArquivos.java' que gera os arquivos de teste;
• adicionado o '.txt'  no .gitignore para ignorar os arquivos de testes;
• no arquivo 'main.c' agora é testado com input de arquivo, ao inves de input manual
• em bTree.c foi adicionada a versão Iterativa da função addNoR (que é recursiva)
• a criação de No agora tem sua própria função 'criaNo'
• em a struct Aluno agora guarda as notas em 'int' (antes era 'float')

+++ gBST/aluno.h
@@ -3,7 +3,7 @@
 
 #include "func/aluno.c"
 
-Aluno *criaAluno(int mat, char *nome, float nota);
+Aluno *criaAluno(int mat, char *nome, int nota);
 int compara(void* a, void* b);
 void imprimeElemento(void* a);
